+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unrelesaed
+
+### Added
+
+- The defualt accuracy and minimum duration of RTK surveys can now be configured
+  in the `rtk` extension.
+
 ## [2.14.0] - 2023-06-16
 
 ### Changed